+++ packages/gamut/src/FlexGrid/index.tsx
@@ -1,3 +1,3 @@
-export { default as Grid } from './Grid';
-export { default as Row } from './Row';
-export { default as Col } from './Col';
+export * from './FlexGrid';
+export * from './Row';
+export * from './Col';