+++ packages/styleguide/stories/Foundations/Layouts/FlexGrid.stories.tsx
@@ -1,4 +1,4 @@
-import { Item, Grid, Row, Col } from '@codecademy/gamut/src';
+import { Item, FlexGrid, Row, Col } from '@codecademy/gamut/src';
 import gamut from '@codecademy/gamut-styles/utils/variables';
 import React from 'react';
 
@@ -17,7 +17,7 @@ const Box = () => (
     style={{
       height: 30,
       marginBottom: '0.5rem',
-      backgroundColor: gamut.deprecatedColors.swatches.ccBlue[500],
+      backgroundColor: gamut.colors.blue[500],
     }}
   />
 );
@@ -25,7 +25,7 @@ const Box = () => (
 const defaultGridProps = {
   style: {
     padding: '1rem',
-    backgroundColor: gamut.deprecatedColors.swatches.ccBlue[100],
+    backgroundColor: gamut.colors.blue[100],
   },
 };
 
@@ -39,7 +39,7 @@ export const fixedGridDefault = decoratedStory('Fixed Grid (Default)', () => (
       This grid is <em>"fixed"</em> meaning that its width is constant: it will
       not shrink to the size of its container.
     </StoryDescription>
-    <Grid {...defaultGridProps}>
+    <FlexGrid {...defaultGridProps}>
       <Row>
         <Col xs={12} sm={3} md={2} lg={1}>
           <Box />
@@ -71,7 +71,7 @@ export const fixedGridDefault = decoratedStory('Fixed Grid (Default)', () => (
           <Box />
         </Col>
       </Row>
-    </Grid>
+    </FlexGrid>
   </StoryTemplate>
 ));
 
@@ -81,7 +81,7 @@ export const fluid = decoratedStory(() => (
       Adding the fluid prop allows the grid to grow and shrink freely inside its
       container.
     </StoryDescription>
-    <Grid {...defaultGridProps} fluid>
+    <FlexGrid {...defaultGridProps} fluid>
       <Row>
         <Col xs={12} sm={3} md={2} lg={1}>
           <Box />
@@ -113,7 +113,7 @@ export const fluid = decoratedStory(() => (
           <Box />
         </Col>
       </Row>
-    </Grid>
+    </FlexGrid>
   </StoryTemplate>
 ));
 
@@ -122,7 +122,7 @@ export const offsets = decoratedStory(() => (
     <StoryDescription>
       A column can be offset with the <code>xs</code> property.
     </StoryDescription>
-    <Grid fluid {...defaultGridProps}>
+    <FlexGrid fluid {...defaultGridProps}>
       <Row>
         <Col xsOffset={11} xs={1}>
           <Box />
@@ -158,7 +158,7 @@ export const offsets = decoratedStory(() => (
           <Box />
         </Col>
       </Row>
-    </Grid>
+    </FlexGrid>
   </StoryTemplate>
 ));
 
@@ -167,7 +167,7 @@ export const centered = decoratedStory(() => (
     <StoryDescription>
       You can center the content of a row by passing it <code>center</code>.
     </StoryDescription>
-    <Grid fluid {...defaultGridProps}>
+    <FlexGrid fluid {...defaultGridProps}>
       <Row>
         <Col xs={12}>
           <Row center="xs">
@@ -177,6 +177,6 @@ export const centered = decoratedStory(() => (
           </Row>
         </Col>
       </Row>
-    </Grid>
+    </FlexGrid>
   </StoryTemplate>
 ));