+++ packages/styleguide/stories/Organisms/GridForm.stories.tsx
@@ -1,4 +1,4 @@
-import { GridForm } from '@codecademy/gamut/src';
+import { GridForm, Input } from '@codecademy/gamut/src';
 import { action } from '@storybook/addon-actions';
 import React from 'react';
 
@@ -58,11 +58,14 @@ export const gridForm = decoratedStory(() => (
           type: 'text',
         },
         {
-          label: 'Simple select',
+          label: 'Simple select (required)',
           name: 'simple-select',
-          options: ['One fish', 'Two fish', 'Red fish', 'Blue fish'],
+          options: ['', 'One fish', 'Two fish', 'Red fish', 'Blue fish'],
           size: 5,
           type: 'select',
+          validation: {
+            required: 'Please select an option',
+          },
         },
         {
           label: 'Upload a cat image (we support pdf, jpeg, or png files)',
@@ -114,6 +117,237 @@ export const gridForm = decoratedStory(() => (
           name: 'enough-swag',
           size: 3,
           type: 'checkbox',
+        },
+        {
+          label: 'End User License Agreement',
+          description:
+            'I accept the terms and conditions (required or else!!!)',
+          name: 'eula-checkbox-required',
+          size: 12,
+          type: 'checkbox',
+          validation: {
+            required: 'Please check the box to agree to the terms.',
+          },
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Submit Me!?',
+        size: 12,
+      }}
+    />
+  </StoryTemplate>
+));
+
+export const gridFormWithSubmitButtonPosition = decoratedStory(() => (
+  <StoryTemplate status={StoryStatus.Ready}>
+    <StoryDescription>
+      We can position the submit button by passing the position prop with a
+      value of left, center or right.
+    </StoryDescription>
+    <GridForm
+      fields={[
+        {
+          label: 'Simple text',
+          name: 'simple-text',
+          size: 12,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Right Submit!?',
+        position: 'right',
+        size: 12,
+      }}
+    />
+    <GridForm
+      fields={[
+        {
+          label: 'Simple text',
+          name: 'simple-text',
+          size: 12,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Center Submit!?',
+        position: 'center',
+        size: 12,
+      }}
+    />
+    <GridForm
+      fields={[
+        {
+          label: 'Simple text',
+          name: 'simple-text',
+          size: 12,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Left Submit!?',
+        position: 'left',
+        size: 12,
+      }}
+    />
+  </StoryTemplate>
+));
+
+export const gridFormWithSubmitButtonColor = decoratedStory(() => (
+  <StoryTemplate status={StoryStatus.Ready}>
+    <StoryDescription>
+      We can specify the color of our button by passing the theme prop with a
+      button atom theme prop value. i.e. 'brand-blue', 'brand-purple'. The
+      default value is 'brand-purple'.
+    </StoryDescription>
+    <GridForm
+      fields={[
+        {
+          label: 'Simple text',
+          name: 'simple-text',
+          size: 12,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Default Purple Submit!?',
+        size: 12,
+      }}
+    />
+    <GridForm
+      fields={[
+        {
+          label: 'Simple text',
+          name: 'simple-text',
+          size: 12,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Blue Submit!?',
+        size: 12,
+        theme: 'brand-blue',
+      }}
+    />
+  </StoryTemplate>
+));
+
+export const gridFormWithInlineSubmitButton = decoratedStory(() => (
+  <StoryTemplate status={StoryStatus.NotReady}>
+    <StoryDescription>
+      We can make the Submit button inline with an input by setting the column
+      sizes so they fit on the same row. e.g size 8 for an input and size 4 for
+      the submit.
+      <br />
+      Caveats: We need to adjust params to allow us to properly align elements
+      for the following use cases.
+    </StoryDescription>
+    <h4>
+      Submit with no label text (but label still rendered) does not align
+      properly with the checkbox
+    </h4>
+    <GridForm
+      fields={[
+        {
+          description: 'I approve of inline swag',
+          name: 'enough-swag',
+          size: 8,
+          type: 'checkbox',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Inline Submit!?',
+        size: 4,
+        position: 'right',
+      }}
+    />
+    <hr />
+    <h4>
+      Submit even when the lable has visible text, is still not aligned right
+    </h4>
+    <GridForm
+      fields={[
+        {
+          label: 'Label',
+          name: 'email',
+          size: 8,
+          type: 'text',
+        },
+      ]}
+      onSubmit={async values => {
+        action('Form Submitted')(values);
+      }}
+      submit={{
+        contents: 'Inline Submit!?',
+        size: 4,
+        position: 'right',
+      }}
+    />
+  </StoryTemplate>
+));
+
+export const gridFormWithCustomInput = decoratedStory(() => (
+  <StoryTemplate status={StoryStatus.Ready}>
+    <StoryDescription>
+      Some forms, such as the checkout flows that use Recurly, need to define
+      their own inputs. We can specify a 'custom' field type to with a{' '}
+      <a
+        href="https://reactjs.org/docs/render-props.html"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        render prop
+      </a>
+      .
+    </StoryDescription>
+
+    <GridForm
+      fields={[
+        {
+          render: ({ className, error, setValue }) => (
+            <div className={className}>
+              <Input
+                error={!!error}
+                id="custom-text-input"
+                type="text"
+                onChange={event => setValue(event.target.value)}
+              />
+              <span aria-label="Dancing person" role="img">
+                🕺
+              </span>
+            </div>
+          ),
+          label: 'Gimme two more swags',
+          name: 'custom-input',
+          validation: {
+            required: true,
+            pattern: {
+              value: /swag(.*)swag/,
+              message: 'Still not enough swag, what are you doing... 💢',
+            },
+          },
+          size: 12,
+          type: 'custom',
         },
       ]}
       onSubmit={async values => {